+++ sys/kern/src/kipc.rs
@@ -364,7 +364,7 @@ fn read_task_dump_region(
             // succeed (see: the comparisons between base+size and region.size
             // above).
             let offset = from.base_addr() - tcb_base;
-            let tcb = &target_task[offset as usize..from.len()];
+            let tcb = &target_task[offset..from.len()];
 
             let to = caller_task
                 .try_write(&mut response)
